ENH: Disambiguate a call to vnl_sqrt

+++ Base/Filtering/itktubePadImageFilter.h
@@ -195,7 +195,7 @@ private:
 
   bool isPrime( int n )
     {
-    int last = (int)vcl_sqrt( n );
+    int last = (int)vcl_sqrt( static_cast<float>( n ) );
     for( int x=2; x<=last; x++ )
       {
       if( n%x == 0 )
